requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pandas
duckdb
pyyaml
openpyxl
plotly
streamlit
